# api/openapi.yaml
openapi: 3.0.3
info:
  title: Inbox Whisperer API
  version: 0.1.0
  description: |
    OpenAPI specification for the Inbox Whisperer backend API.
    This spec is always kept in sync with implemented endpoints.
servers:
  - url: http://localhost:8080
paths:
  /api/auth/login:
    get:
      tags: [Auth]
      summary: Start Google OAuth2 login
      description: Redirects the user to Google's OAuth2 consent screen. Sets up session state for CSRF protection.
      responses:
        '302':
          description: Redirect to Google OAuth2
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/callback:
    get:
      tags: [Auth]
      summary: OAuth2 callback handler
      description: Handles the Google OAuth2 redirect. Exchanges code for token, creates user if needed, and sets session.
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to frontend on success
        '400':
          description: Invalid or missing state/code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/emails:
    get:
      tags: [Email]
      summary: Fetch user's emails
      description: Fetches the latest emails for the authenticated user and returns a list of email summaries.
      responses:
        '200':
          description: List of email summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmailSummary'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/emails/{id}:
    get:
      tags: [Email]
      summary: Get full email content
      description: Returns the full content of a specific email by ID for the authenticated user.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailContent'
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Email]
      summary: Update email (not implemented)
      description: Not implemented yet.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '501':
          description: Not implemented
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Email]
      summary: Delete email (not implemented)
      description: Not implemented yet.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '501':
          description: Not implemented
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    EmailSummary:
      type: object
      properties:
        id:
          type: string
          example: 1789a2b1cdefg
        subject:
          type: string
          example: "Welcome to Inbox Whisperer!"
        from:
          type: string
          example: "notifications@example.com"
        snippet:
          type: string
          example: "This is a preview of your email..."
        date:
          type: string
          format: date-time
          example: 2025-04-22T00:00:00Z
    EmailContent:
      type: object
      properties:
        id:
          type: string
          example: 1789a2b1cdefg
        subject:
          type: string
          example: "Welcome to Inbox Whisperer!"
        from:
          type: string
          example: "notifications@example.com"
        to:
          type: string
          example: "user@example.com"
        date:
          type: string
          format: date-time
          example: 2025-04-22T00:00:00Z
        body:
          type: string
          example: "Hello and welcome..."
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: User not found